# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ISP Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Customer self-registration (after OTP verification)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone + OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/claim-account:
    post:
      operationId: claimAccount
      tags: [auth]
      summary: Imported customer claims their account with OTP + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimAccountInput"
      responses:
        "200":
          description: Account claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all internet packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Admin creates a new package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /packages/{id}:
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Admin updates a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Admin deletes a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CUSTOMERS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, phone, name, role, status, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, customer]
        status:
          type: string
          enum: [active, suspended, pending-claim]
        address:
          type: ["string", "null"]
        zone:
          type: ["string", "null"]
        createdAt:
          type: string
    SendOtpInput:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
    RegisterInput:
      type: object
      required: [phone, otp, name, password]
      properties:
        phone:
          type: string
        otp:
          type: string
        name:
          type: string
        password:
          type: string
        address:
          type: string
        zone:
          type: string
    LoginInput:
      type: object
      required: [phone, otp]
      properties:
        phone:
          type: string
        otp:
          type: string
    ClaimAccountInput:
      type: object
      required: [phone, otp, password]
      properties:
        phone:
          type: string
        otp:
          type: string
        password:
          type: string
    Package:
      type: object
      required: [id, name, speedMbps, price, validity, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        speedMbps:
          type: integer
        price:
          type: number
        validity:
          type: string
          enum: [monthly, quarterly, yearly]
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    PackageInput:
      type: object
      required: [name, speedMbps, price, validity]
      properties:
        name:
          type: string
        speedMbps:
          type: integer
        price:
          type: number
        validity:
          type: string
          enum: [monthly, quarterly, yearly]
        description:
          type: string
        isActive:
          type: boolean
    PackageUpdate:
      type: object
      properties:
        name:
          type: string
        speedMbps:
          type: integer
        price:
          type: number
        validity:
          type: string
          enum: [monthly, quarterly, yearly]
        description:
          type: string
        isActive:
          type: boolean
...